Added test cases for LearnRORg...running to make sure they all pass.
