+++ project/crud-microservice-app/golang/testing-project/server.go
@@ -27,6 +27,7 @@ func handleRequests() {
 	myRouter.HandleFunc("/getLearnOrg", getLearnOrg).Methods("POST")       //Get LearnROrg
 	//Serve our validation APIs
 	myRouter.HandleFunc("/giveAllUsernames", giveAllUsernames).Methods("GET")       //Get all our Usernames
+	myRouter.HandleFunc("/giveAllLearnROrg", giveAllLearnROrg).Methods("GET")       //Get all our LearnROrg Names
 	myRouter.HandleFunc("/randomIDCreationAPI", randomIDCreationAPI).Methods("GET") //Get a random ID
 	myRouter.HandleFunc("/userLogin", userLogin).Methods("POST")                    //Get a random ID
 	//Serve our static files
