+++ project/webapp/golang/testing-project/logging/weblog.txt
@@ -16,3 +16,5 @@
 2021/06/10 15:14:48 User found
 2021/06/10 15:14:48 User found
 2021/06/10 15:18:33 No usernames returned...this could be the site's first deployment with no users!
+2021/06/12 13:10:57 This is a test message
+2021/06/12 13:20:42 This is a test message
